Update variable of model

See merge request !341

+++ src/main/java/utilities/model/dashboard/orders/orderdetail/OrderInfo.java
@@ -27,7 +27,7 @@ public class OrderInfo {
     private double totalDiscount;
     private List<Object> discounts;
     private String transactionNumber;
-    private int totalTaxAmount;
+    private double totalTaxAmount;
     private double loyaltyPoint;
     private int pointAmount;
     private double receivedAmount;
